+++ frontend/src/components/popup/createCardForm/CardForm.tsx
@@ -27,6 +27,8 @@ const CardForm = ({ categoryId, categoryType, maxIndex, boardColor }: CardFormPr
   const [accessToken] = useAtom(accessTokenAtom);
   const [open, setOpen] = useState(false);
   const [columns, setColumns] = useAtom(columnsAtom);
+  const [loading, setLoading] = useState(false);
+
   // Job information
   const [companyName, setCompanyName] = useState('');
   const [companyIndustry, setCompanyIndustry] = useState('');
@@ -46,6 +48,8 @@ const CardForm = ({ categoryId, categoryType, maxIndex, boardColor }: CardFormPr
   const [flowProcesses, setFlowProcesses] = useState<FlowForm[]>([]);
 
   const handleSaveCard = async () => {
+    setLoading(true);
+
     const cardDetail: FullJobCreate = {
       job: {
         category_id: categoryId,
@@ -78,6 +82,7 @@ const CardForm = ({ categoryId, categoryType, maxIndex, boardColor }: CardFormPr
       category.jobs.sort((a, b) => a.card_position - b.card_position);
     });
     setColumns(newColumns);
+    setLoading(false);
     handleClose();
   };
 
@@ -164,7 +169,7 @@ const CardForm = ({ categoryId, categoryType, maxIndex, boardColor }: CardFormPr
           </DialogContent>
           <DialogActions>
             <Button onClick={handleClose}>キャンセル</Button>
-            <Button type="submit">保存</Button>
+            <Button type="submit">{loading ? '保存中...' : '保存'}</Button>
           </DialogActions>
         </form>
       </Dialog>
